Omae wa mou shindeiru'ed aws.py
