+++ ScoutSuite/providers/aws/provider.py
@@ -5,13 +5,12 @@ import os
 import boto3
 
 from ScoutSuite.core.console import print_debug, print_error, print_exception, print_info
-from ScoutSuite.providers.aws.aws import get_aws_account_id
 from ScoutSuite.providers.aws.configs.services import AWSServicesConfig
 from ScoutSuite.providers.aws.services.vpc import put_cidr_name
 from ScoutSuite.providers.base.configs.browser import combine_paths, get_object_at, get_value_at
 from ScoutSuite.providers.base.provider import BaseProvider
 from ScoutSuite.utils import manage_dictionary
-from ScoutSuite.providers.aws.utils import ec2_classic
+from ScoutSuite.providers.aws.utils import ec2_classic, get_aws_account_id
 
 
 class AWSProvider(BaseProvider):
