+++ ScoutSuite/providers/aws/services/ec2.py
@@ -9,12 +9,11 @@ import base64
 import netaddr
 
 from ScoutSuite.core.console import print_exception, print_info
-from ScoutSuite.providers.aws.aws import get_name
 from ScoutSuite.providers.aws.configs.regions import RegionalServiceConfig, RegionConfig, api_clients
 from ScoutSuite.providers.aws.configs.vpc import VPCConfig
 from ScoutSuite.providers.base.configs.browser import get_attribute_at
 from ScoutSuite.utils import manage_dictionary
-from ScoutSuite.providers.aws.utils import ec2_classic, get_keys
+from ScoutSuite.providers.aws.utils import ec2_classic, get_keys, get_name
 from ScoutSuite.core.fs import load_data
 
 ########################################
