+++ ScoutSuite/providers/aws/services/rds.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
 
 from ScoutSuite.core.console import print_error, print_exception
-from ScoutSuite.providers.aws.aws import handle_truncated_response
 from ScoutSuite.providers.aws.configs.regions import RegionalServiceConfig, RegionConfig, api_clients
 from ScoutSuite.providers.aws.configs.vpc import VPCConfig
 from ScoutSuite.utils import manage_dictionary
-from ScoutSuite.providers.aws.utils import ec2_classic
+from ScoutSuite.providers.aws.utils import ec2_classic, handle_truncated_response
 
 
 ########################################
