+++ ScoutSuite/providers/base/configs/services.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from ScoutSuite.core.console import print_error, print_exception, print_debug
-from ScoutSuite.providers.aws.aws import get_partition_name
+from ScoutSuite.providers.aws.utils import get_partition_name
 
 
 class BaseServicesConfig(object):
